% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{gb.gene.cor}
\alias{gb.gene.cor}
\title{Get gene coordinates}
\usage{
gb.gene.cor(gb)
}
\arguments{
\item{gb}{gb file to get the genes from.}
}
\value{
he crude table extracted from gb file for genes.
}
\description{
Get gene coordinates
}
