% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{gene.cordinates}
\alias{gene.cordinates}
\title{Get genes coordinates}
\usage{
gene.cordinates(gb)
}
\arguments{
\item{gb}{gb file to get the genes from.}
}
\value{
hpolished table of genes and their coordinates, with the gene names,
start, end of first part and the second parts (first to fifth columns).
The orders are reflected with the swapping of values.
}
\description{
Get genes coordinates
}
