% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{gene.name}
\alias{gene.name}
\title{Get gene names}
\usage{
gene.name(gb, type)
}
\arguments{
\item{gb}{gb file to get the names from.}

\item{type}{either gene, tRNA, mRNA or rRNA.}
}
\value{
vector with the gene names.
}
\description{
Get gene names
}
