% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{get_data_from_gb}
\alias{get_data_from_gb}
\title{Getting data from gb file}
\usage{
get_data_from_gb(gbfile, local.file = FALSE)
}
\arguments{
\item{gbfile}{A character. It is the GI or GenBank accession for the
interested specie's chloroplast genome files.}

\item{local.file}{A logical value. If it is \code{TRUE}, the value passed to
\code{gbfile} should be a local GB file.}
}
\value{
list with:
\itemize{
  \item gb: read gb file.
  \item genome: A DNAString object. It is a class defined in \code{Biostrings}
   package. It contains the whole genome sequence of a specie from 
   the given gb file.
  \item l: length of the genome.
  \item gene_table: table with the genes and positions.
}
}
\description{
Extracting the needed information from a gb file.
}
