% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{get_genome}
\alias{get_genome}
\title{# Questions:
# Is it necessary we check column number of ORIGIN field in function "get_genome"?}
\usage{
get_genome(gb)
}
\arguments{
\item{gb}{vector of characters. Genebank file returned by
\code{\link{fetch.gb}} or \code{\link{read.gb}} function.}
}
\value{
The genome of the sequence that is deposited at the end of the GeneBank file in DNAString object format
}
\description{
Genome extracter
}
\details{
Extracting the fasta format chloroplast genome from a GeneBank File
}
