% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ir.R
\name{get_ir}
\alias{get_ir}
\title{Generate information for the IR region.}
\usage{
get_ir(genome)
}
\arguments{
\item{genome}{A DNAString object. It is a class defined in \code{Biostrings}
package. It contains the whole genome sequence of a specie.}

\item{l}{A number. Indicates the genome length.}
}
\value{
A named list with:
\itemize{
  \item ir_table: A vector containing IR region information:
  \itemize{
     \item IRA_start: gene where IRA starts.
     \item IRB_start: gene where IRB starts.
     \item IR_len: length of the inverted repeats.
     \item gene_len: length of the genome.
  }
  \item indel_table: a dataframe with the mismatches found with these columns:
  \itemize{
     \item mismatch_type: type of the mismatch.
     \item position: where in the genome it is.
     \item string: letter changed.
     \item col: color.
  }
}
}
\description{
Generate information for the IR region.
}
