% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ir.R
\name{get_ir_positions}
\alias{get_ir_positions}
\title{Transforms ir_table to a matrix.}
\usage{
get_ir_positions(ir_table, genome_length)
}
\arguments{
\item{ir_table}{dataframe with the ir info obtained with \code{\link{get_ir}}.}

\item{genome_length}{A number. Indicates the genome length.}
}
\value{
A list with:
\itemize{
   \item IRA_start: gene where IRA starts.
   \item IRB_start: gene where IRB starts.
   \item IRA_len: length of the inverted repeats A.
   \item IRB_len: length of the inverted repeats B.
   \item gene_len: length of the genome.
}
}
\description{
Transforms ir_table to a matrix.
}
