% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{get_pc_J}
\alias{get_pc_J}
\title{get_pc_J}
\usage{
get_pc_J(J.pos, jlens, ir_info)
}
\arguments{
\item{J.pos}{The position of the J site: JLB, JSB, JSA, and JLA for 1,2,3, 
and 4 respectively.}

\item{jlens}{named list with the distance in the plot for the junction sites, 
in the same order as J.pos: jlb.len, jsb.len, jsa.len, jla.len.}

\item{ir_info}{ir information for the genome}
}
\value{
list with pc (position on the plot for that junction) and J (position
position on the genome of the junction)
}
\description{
Get pc and J values auxiliar function
}
