% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{GnlBuilder}
\alias{GnlBuilder}
\title{sp.name <- function(definition){
  # if (text){
    # sp <- gsub("(DEFINITION\\ \\ )", "", definition[2], perl = TRUE)
    # sp <- sub("(\\w+\\s+\\w+).*", "\\1", sp, perl = TRUE)
  # } else {
    sp <- sub("(\\w+\\s+\\w+).*", "\\1", definition, perl = TRUE)
  # }
  return(sp)
}}
\usage{
GnlBuilder(dogma)
}
\arguments{
\item{dogma}{a character. It contains the the read dogma file.}
}
\value{
a list with the genes.
}
\description{
Transform the dogma file gene list to one suited for the functions.
}
