% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ir.R
\name{irDetect}
\alias{irDetect}
\title{Generate information for the IR region.}
\usage{
irDetect(genome, seed.size = 1000)
}
\arguments{
\item{genome}{A DNAString object. It is a class defined in \code{Biostrings}
package. It contains the whole genome sequence of a specie.}

\item{seed.size}{Seed to set the starting point.}
}
\value{
A named list with:
\itemize{
  \item ir_table: A dataframe containing IR region information.
  \item indel_table: a dataframe with the mismatches found with these columns:
  \itemize{
     \item mismatch_type: type of the mismatch.
     \item position: where in the genome it is.
     \item string: letter changed.
     \item col: color.
  }
}
}
\description{
Generate information for the IR region.
}
