% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{IRs}
\alias{IRs}
\title{Generates all needed information to plot the IR region given gbfiles}
\usage{
IRs(gbfiles, Sfiles, progress)
}
\arguments{
\item{gbfiles}{named list with the gbfiles. The parameters of the list are:
\itemize{
  \item gbfile: string with the accession number os path with the gbfile.
  \item local.file: boolean TRUE if it is a pth with a local gbfile or 
        FALSE if accession number.
}}

\item{Sfiles}{list of booleans. Indicates if the file needs a SSC reversion}

\item{progress}{progress bar from shiny to show how many species have been 
analyzed so far.}
}
\value{
None. The variables set here are used as global variables (l, IRList, 
        IndelList, FasList, nuclw, spnames, GeneList)
}
\description{
Generates all needed information to plot the IR region given gbfiles
}
