% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{IRsD}
\alias{IRsD}
\title{Generates all needed information to plot the IR region given fasta and gene 
info files}
\usage{
IRsD(dgfiles, fastafiles, irfiles, nfiles, progress)
}
\arguments{
\item{dgfiles}{list with the dogma files (files with gene info) paths.}

\item{fastafiles}{list with the fasta files paths.}

\item{irfiles}{list with ir information provided.}

\item{nfiles}{list of species names.}

\item{progress}{progress bar from shiny to show how many species have been 
analyzed so far.}
}
\value{
None. The variables set here are used as global variables (IRListDinp, 
        IndelList, FasList, GeneList, nuclw, spnames)
}
\description{
Generates all needed information to plot the IR region given fasta and gene 
info files
}
