% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{IRsD2}
\alias{IRsD2}
\title{Plotting data function for dogma and fasta files}
\usage{
IRsD2(file = "IR_out", theme, sample = FALSE)
}
\arguments{
\item{file}{filename where the plot is done.}

\item{theme}{dataframe with the colors chosen on the web.}

\item{sample}{boolean. TRUE if the plot is just a sample (2 species), FALSE
if all the species are to be plotted.}
}
\value{
None.
}
\description{
Creates the entire plot.
}
