% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{JunctRadiusGeneFinder}
\alias{JunctRadiusGeneFinder}
\title{Find the genes near the junction coordinate based on the output of 
the \code{\link{gene.cordinate}} function.}
\usage{
JunctRadiusGeneFinder(gene.cordinates, IRinfo, J.pos, radius, silence = TRUE)
}
\arguments{
\item{gene.cordinates}{cordinates given by the \code{\link{gene.cordinate}} 
function.}

\item{J.pos}{The position of the J site: JLB, JSB, JSA, and JLA for 1,2,3, 
and 4 respectively}

\item{radius}{number with the radius wanted}

\item{silence}{boolean}

\item{Irinfo}{vector with ira start, irb start, ir length and genome length.}
}
\value{
gene list
}
\description{
Find the genes near the junction coordinate based on the output of 
the \code{\link{gene.cordinate}} function.
}
