% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{JunctRadiusIndelFinder}
\alias{JunctRadiusIndelFinder}
\title{Find the mismatches near the junction cordinate based on the indel_table 
given by \code{\link{get_ir}} function.}
\usage{
JunctRadiusIndelFinder(indel_table, IRinfo, J.pos, radius)
}
\arguments{
\item{indel_table}{dataframe with mismatches given by \code{\link{get_ir}} 
function.}

\item{IRinfo}{vector with ira start, irb start, ira length, irb length and genome length.}

\item{J.pos}{The position of the J site: JLB, JSB, JSA, and JLA for 1,2,3, 
and 4 respectively}

\item{radius}{number with the radius wanted}
}
\value{
list with indel inside radius
}
\description{
Find the mismatches near the junction cordinate based on the indel_table 
given by \code{\link{get_ir}} function.
}
