% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{LSC}
\alias{LSC}
\title{LSC place}
\usage{
LSC(IRinfo)
}
\arguments{
\item{IRinfo}{vector with ira start, irb start, ira length, irb length and 
genome length.}
}
\value{
number: place where LSC starts.
}
\description{
Gives where LSC starts
}
