% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{LSCDinp}
\alias{LSCDinp}
\title{LSC place}
\usage{
LSCDinp(IRinp)
}
\arguments{
\item{IRinfo}{vector with ira start, irb start, ir length and genome length.}
}
\value{
number: place where LSC starts.
}
\description{
Gives where LSC starts
}
