% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{Max.Radius}
\alias{Max.Radius}
\title{Max radius finder}
\usage{
Max.Radius(J.pos, l, genelist, irlist)
}
\arguments{
\item{J.pos}{The position of the J site: JLB, JSB, JSA, and JLA for 1,2,3, 
and 4 respectively.}

\item{l}{number of species to plot.}

\item{genelist}{list with the genes.}

\item{IRlist}{vector with ira start, irb start, ira length, irb length and genome length.}
}
\value{
list with the radius for each species (the same if they are similar,
different if they are not similar species).
}
\description{
Finding the best radius for the vicinity of a junction site. This radius 
determines which genes appear on the plot. The radius is decided according to
when a gene can be plotted. If the species are similar, the radius chosen 
will be the same for all of them. Else, each species will have its own radius 
and the plot won't be in scale.
}
