% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{n_int_digits}
\alias{n_int_digits}
\title{Number character count}
\usage{
n_int_digits(x)
}
\arguments{
\item{x}{number}
}
\value{
number of characters
}
\description{
Counts how many characters a number has
}
