% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot.data.aux}
\alias{plot.data.aux}
\title{Plotting data auxiliar function}
\usage{
\method{plot}{data.aux}(radius_list, J.pos, l, jlens, theme)
}
\arguments{
\item{radius_list}{list with the calculated radius for the species.}

\item{J.pos}{The position of the J site: JLB, JSB, JSA, and JLA for 1,2,3, 
and 4 respectively.}

\item{l}{number of species.}

\item{jlens}{named list with the distance in the plot for the junction sites, 
in the same order as J.pos: jlb.len, jsb.len, jsa.len, jla.len.}

\item{theme}{dataframe with the colors chosen on the web.}
}
\value{
None.
}
\description{
Calls all the needed functions to plot the information. It dies it for all
the species.
}
