% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{sp.name}
\alias{sp.name}
\title{rdnFixer<- function(genome){
  seq <- Biostrings::toString(genome)
  seq <- unlist(strsplit(seq, ""))
  seq[which(seq=="U")]<-sample(c("T"), length(which(seq=="U")), TRUE)
  seq[which(seq=="R")]<-sample(c("A", "G"), length(which(seq=="R")), TRUE)
  seq[which(seq=="Y")]<-sample(c("C", "T"), length(which(seq=="Y")), TRUE)
  seq[which(seq=="S")]<-sample(c("C", "G"), length(which(seq=="S")), TRUE)
  seq[which(seq=="W")]<-sample(c("A", "T"), length(which(seq=="W")), TRUE)
  seq[which(seq=="K")]<-sample(c("G", "T"), length(which(seq=="K")), TRUE)
  seq[which(seq=="M")]<-sample(c("C", "A"), length(which(seq=="M")), TRUE)
  seq[which(seq=="B")]<-sample(c("C", "G", "T"), length(which(seq=="B")), TRUE)
  seq[which(seq=="D")]<-sample(c("A", "G", "T"), length(which(seq=="D")), TRUE)
  seq[which(seq=="H")]<-sample(c("C", "A", "T"), length(which(seq=="H")), TRUE)
  seq[which(seq=="V")]<-sample(c("C", "A", "G"), length(which(seq=="V")), TRUE)
  seq[which(seq=="N")]<-sample(c("C", "G", "T", "A"), length(which(seq=="N")), TRUE)
  seq[which(seq=="-")]<-sample(c("C", "G", "T", "A"), length(which(seq=="-")), TRUE)
  seq <- paste(seq, collapse = "")
  seq <- Biostrings::DNAString(seq)
  return(seq)
}
Extract specie's name from GB file}
\usage{
sp.name(gb)
}
\arguments{
\item{gb}{a character. It contains the the read GenBank file.}
}
\value{
a character contains the specie's name.
}
\description{
rdnFixer<- function(genome){
  seq <- Biostrings::toString(genome)
  seq <- unlist(strsplit(seq, ""))
  seq[which(seq=="U")]<-sample(c("T"), length(which(seq=="U")), TRUE)
  seq[which(seq=="R")]<-sample(c("A", "G"), length(which(seq=="R")), TRUE)
  seq[which(seq=="Y")]<-sample(c("C", "T"), length(which(seq=="Y")), TRUE)
  seq[which(seq=="S")]<-sample(c("C", "G"), length(which(seq=="S")), TRUE)
  seq[which(seq=="W")]<-sample(c("A", "T"), length(which(seq=="W")), TRUE)
  seq[which(seq=="K")]<-sample(c("G", "T"), length(which(seq=="K")), TRUE)
  seq[which(seq=="M")]<-sample(c("C", "A"), length(which(seq=="M")), TRUE)
  seq[which(seq=="B")]<-sample(c("C", "G", "T"), length(which(seq=="B")), TRUE)
  seq[which(seq=="D")]<-sample(c("A", "G", "T"), length(which(seq=="D")), TRUE)
  seq[which(seq=="H")]<-sample(c("C", "A", "T"), length(which(seq=="H")), TRUE)
  seq[which(seq=="V")]<-sample(c("C", "A", "G"), length(which(seq=="V")), TRUE)
  seq[which(seq=="N")]<-sample(c("C", "G", "T", "A"), length(which(seq=="N")), TRUE)
  seq[which(seq=="-")]<-sample(c("C", "G", "T", "A"), length(which(seq=="-")), TRUE)
  seq <- paste(seq, collapse = "")
  seq <- Biostrings::DNAString(seq)
  return(seq)
}
Extract specie's name from GB file
}
