% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{SSC}
\alias{SSC}
\title{SSC place}
\usage{
SSC(IRinfo)
}
\arguments{
\item{IRinfo}{vector with ira start, irb start, ira length, irb length and 
genome length.}
}
\value{
number: place where SSC starts.
}
\description{
Gives where SSC starts
}
