% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{SSCDinp}
\alias{SSCDinp}
\title{SSC place}
\usage{
SSCDinp(IRinp)
}
\arguments{
\item{IRinfo}{vector with ira start, irb start, ir length and genome length.}
}
\value{
number: place where SSC starts.
}
\description{
Gives where SSC starts
}
