% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{SSCrev}
\alias{SSCrev}
\title{Reverse SSC section in the plot.}
\usage{
SSCrev(g, SSC_start, SSC_end)
}
\arguments{
\item{g}{gene coordinates for SSC region.}

\item{SSC_start}{number: SSC start}

\item{SSC_end}{number: SSC end}
}
\value{
g edited to be reversed.
}
\description{
Reverse SSC section in the plot.
}
