% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{toIRDinp}
\alias{toIRDinp}
\title{Transform IRInfo format to the one used in dogma functions}
\usage{
toIRDinp(IRinfo)
}
\arguments{
\item{IRinfo}{vector with ira ini, ira end, len ir, len genome.}
}
\value{
vector with ira ini, ira end, irb ini, irb end, len genome
}
\description{
Transform IRInfo format to the one used in dogma functions
}
