% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{toIRinfo}
\alias{toIRinfo}
\title{Transform IRInfo format to the one used in gbfiles functions}
\usage{
toIRinfo(IRDinp)
}
\arguments{
\item{IRDinp}{vector with ira ini, ira end, irb ini, irb end, len genome.}
}
\value{
vector with ira ini, ira end, len ir, len genome.
}
\description{
Transform IRInfo format to the one used in gbfiles functions
}
