% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{trnDogma}
\alias{trnDogma}
\title{Transform the dogma file gene list to one suited for the functions with the
gene namesfixed.}
\usage{
trnDogma(dogma)
}
\arguments{
\item{dogma}{a character. It contains the the read dogma file.}
}
\value{
a list with the genes with the gene names fixed.
}
\description{
Transform the dogma file gene list to one suited for the functions with the
gene namesfixed.
}
