% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{SSCrev}
\alias{SSCrev}
\title{Reverse SSC section in the plot}
\usage{
SSCrev(genecord, SSCs, SSCe)
}
\arguments{
\item{genecord}{gene coordinates for SSC region.}

\item{SSCe}{number: SSC end}

\item{SScs}{number: SSC ini}
}
\value{
genecord edited to be reversed.
}
\description{
Reverse SSC section in the plot
}
